ButtPromptApi/requirements.txt
absl-py==2.0.0
accelerate==0.23.0
aiohttp==3.8.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.7.1
astunparse==1.6.3
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
audioread==3.0.0
backoff==2.2.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
datasets==2.14.5
decorator==5.1.1
diffusers==0.19.3
dill==0.3.7
exceptiongroup==1.1.3
fastapi==0.103.1
fastjsonschema==2.18.0
filelock==3.12.4
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.6.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gql==3.4.1
graphql-core==3.2.3
grpcio==1.58.0
h11==0.14.0
h5py==3.9.0
hf-doc-builder==0.4.0
huggingface-hub==0.17.2
idna==3.4
importlib-metadata==6.8.0
invisible-watermark==0.2.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_core==5.3.1
keras==2.13.1
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
lightning-utilities==0.9.0
llvmlite==0.41.0
Markdown==3.4.4
MarkupSafe==2.1.3
mpmath==1.3.0
msgpack==1.0.6
multidict==6.0.4
multiprocess==0.70.15
nbformat==5.9.2
networkx==3.1
numba==0.58.0
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
platformdirs==3.10.0
pooch==1.7.0
protobuf==4.24.3
psutil==5.9.5
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==2.0.9
pytz==2023.3.post1
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rpds-py==0.10.3
rsa==4.9
safetensors==0.3.3
scikit-learn==1.3.1
scipy==1.11.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.6
starlette==0.27.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1+cu118
torchaudio==2.0.2
torchmetrics==1.2.0
torchvision==0.15.2
tqdm==4.66.1
traitlets==5.10.0
transformers==4.33.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.2
websockets==11.0.3
Werkzeug==2.3.7
wrapt==1.15.0
xformers==0.0.21
xxhash==3.3.0
yarl==1.9.2
zipp==3.17.0
websocket-client==1.6.3
signalrcore==0.9.5
